feat: create exception handler - delete method

+++ src/main/java/com/learningjava/course/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.learningjava.course.entities.User;
 import com.learningjava.course.repositories.UserRepository;
+import com.learningjava.course.services.exceptions.DatabaseException;
 import com.learningjava.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -29,11 +32,21 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            if (!repository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+            repository.deleteById(id);
+        } catch (ResourceNotFoundException e) {
+            throw new ResourceNotFoundException(e.getMessage());
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
-        // getReferenceById não busca as informações no db. Ele prepara o obj e depois envia para o banco de dados
+        // getReferenceById não busca as informações no db. Ele prepara o obj e depois
+        // envia para o banco de dados
         User entity = repository.getReferenceById(id);
         updateData(entity, obj);
         return repository.save(entity);
